feat: 주문내역 상세조회

+++ src/main/java/com/hyundai/domain/orders/dto/OrderResponseDto.java
@@ -19,7 +19,7 @@ public class OrderResponseDto {
     private String receiverName;
     private String phoneNumber;
     private String orderStatus;
-    private int orderNumber;
+    private Long orderNumber;
     private String orderRequired;
     private String paymentMethod;
     private List<OrderProductDto> orderProductDtoList;
